Move playwright tests to use project references
