+++ addons/addon-clipboard/test/tsconfig.json
@@ -20,8 +20,7 @@
     },
     "strict": true,
     "types": [
-      "../../../node_modules/@types/node",
-      "../../../out-test/playwright/TestUtils"
+      "../../../node_modules/@types/node"
     ]
   },
   "include": [
@@ -34,6 +33,9 @@
     },
     {
       "path": "../../../src/browser"
+    },
+    {
+      "path": "../../../test/playwright"
     }
   ]
 }
