+++ addons/addon-webgl/test/tsconfig.json
@@ -21,9 +21,7 @@
     "strict": true,
     "types": [
       "../../../node_modules/@types/node",
-      "../../../node_modules/@lunapaint/png-codec",
-      "../../../out-test/playwright/TestUtils",
-      "../../../out-test/playwright/SharedRendererTests"
+      "../../../node_modules/@lunapaint/png-codec"
     ]
   },
   "include": [
@@ -36,6 +34,9 @@
     },
     {
       "path": "../../../src/browser"
+    },
+    {
+      "path": "../../../test/playwright"
     }
   ]
 }
