+++ bin/esbuild.mjs
@@ -92,10 +92,10 @@ let outConfig = {
 let skipOut = false;
 
 /** @type {esbuild.BuildOptions} */
-// let outTestConfig = {
-//   format: 'cjs'
-// }
-// let skipOutTest = false;
+let outTestConfig = {
+  format: 'cjs'
+}
+let skipOutTest = false;
 
 if (config.addon) {
   bundleConfig = {
@@ -108,14 +108,15 @@ if (config.addon) {
     entryPoints: [`addons/addon-${config.addon}/src/**/*.ts`],
     outdir: `addons/addon-${config.addon}/out-esbuild/`
   };
-  // outTestConfig = {
-  //   ...outConfig,
-  //   entryPoints: [`addons/addon-${config.addon}/test/**/*.ts`],
-  //   outdir: `addons/addon-${config.addon}/out-esbuild-test/`
-  // };
+  outTestConfig = {
+    ...outConfig,
+    entryPoints: [`addons/addon-${config.addon}/test/**/*.ts`],
+    outdir: `addons/addon-${config.addon}/out-esbuild-test/`
+  };
 
   if (config.addon === 'ligatures') {
     bundleConfig.platform = 'node';
+    skipOutTest = true;
   }
 
   if (config.addon === 'serialize') {
@@ -173,11 +174,11 @@ if (config.addon) {
     entryPoints: ['src/**/*.ts'],
     outdir: 'out-esbuild/'
   };
-  // outTestConfig = {
-  //   ...outConfig,
-  //   entryPoints: ['test/**/*.ts'],
-  //   outdir: 'out-esbuild-test/'
-  // };
+  outTestConfig = {
+    ...outConfig,
+    entryPoints: ['test/**/*.ts'],
+    outdir: 'out-esbuild-test/'
+  };
 }
 
 if (config.isWatch) {
@@ -185,15 +186,15 @@ if (config.isWatch) {
   if (!skipOut) {
     context(outConfig).then(e => e.watch());
   }
-  // if (!skipOutTest) {
-  //   context(outTestConfig).then(e => e.watch());
-  // }
+  if (!skipOutTest) {
+    context(outTestConfig).then(e => e.watch());
+  }
 } else {
   await build(bundleConfig);
   if (!skipOut) {
     await build(outConfig);
   }
-  // if (!skipOutTest) {
-  //   await build(outTestConfig);
-  // }
+  if (!skipOutTest) {
+    await build(outTestConfig);
+  }
 }
