+++ test/playwright/tsconfig.json
@@ -26,6 +26,7 @@
     },
     "target": "ESNext",
     "module": "commonjs",
+    "composite": true
   },
   "include": [
     "./**/*",
